[Vim] Change mapping to open command mode.
The previous normal mode mapping to open command mode (<leader><leader>)
was giving some problems with easymotion plugin, so I decided to replace
it with Spanish acute accent. This key was previously remapped to <esc>
in all other modes (including command mode), so it can be used for both
opening and closing command mode.