[Vim] Fix mappings to close command mode.
In a previous commit I remapped the Spanish acute accent key to <esc>
in normal, visual and command mode. However, in this last mode, it
executed the command instead of cancelling it. I solved this problem by
making the mapping to move to the start of the command and insert an "
character in order to transform it in a comment and exit without
executing it. I have also made the mapping silent to hide that
character.