[(N)vim] Restructure Vim configuration.
Restructure Vim configuration in order to make it compatible with
Neovim. The purpose of this change is to create a Neovim configuration
that reuses some old Vim configuration. For this I decided to split all
my Vim configuration into three parts:

- Common configuration: This is the configuration that I want to be used
  in both editors.
- Vim-only configuration: This is the configuration that I want to be
  used only in vim.
- Neovim-only configuration: This is the configuration that I want to be
  used only in Neovim, and it will be written mainly in Lua.

Also, I have spit all my plugin configuration into different files for
each one and distributed them depending on which editor I want them to
be loaded on. In Vim, all plugins will still be managed with vim-plug,
whereas in Neovim they will be managed with packer.nvim.