[Nvim] Fix space delay in insert mode.
There was a problem with toggleterm.nvim that was causing a small delay
when inserting a space character in insert mode, making it slower to
show the code completion menu. This was because the command to open the
terminal was mapped to '<space>tt' on both normal and insert mode, so
Neovim waits to see if the user types the another two characters before
inserting a space.

I fixed this by remapping it to another combination that makes use of
the control key, which works better.