[Nvim] Fix language servers not loading.
There was an issue with Lazy.nvim that was causing nvim-lspconfig to no
be loaded properly, making language servers not attach to buffers
automatically in some situations. I think the issue was related with
lazy loading, so I solved it by giving nvim-lspconfig a higher priority
than the default (50), making sure that it loads before all other
plugins.